Got the language toggle working, more work is needed.
